test(users): moved user class to user module
